Updated alt and title in logos in footer , and added vCard for anchor text Aman Grewal
1) Added alt and title attributes to freeCodeCamp and CodePen logo.

It's all about adding value to the site for the user not manipulating the search engines. So, use alt tag to describe the item in detail and the title to provide extra details. Do not have the same text in both as they have different purposes.

2) Updated the link for anchor text 'Aman Grewal'  to vCard

So now my name 'Aman Grewal' in footer is giving the option to store contact information.
The file for this vCard named 'contactAman.vcf' is added for this purpose.
:)